fix: Prevent trailing white space in error codes
Some (maybe all) error messages comes with a white space before the explication of the message. The regex strips everything after `:` but not the space before it.
One example of this is the "Too many attemps" error that yields the following:

> "TOO_MANY_ATTEMPTS_TRY_LATER : Too many unsuccessful login attempts. Please try again later."

Applying your regexp it returns: "TOO_MANY_ATTEMPTS_TRY_LATER ".

I'm not sure what's more efficient between altering the regexp to handle that case or using `.trim()` on the final message.